Moved several functions out of the main utils.ts file (#176)
Closes GSI-627
