+++ src/components/browse/dataset/datasetHeader.tsx
@@ -3,8 +3,8 @@ import { Badge, Row, Col, CloseButton } from "react-bootstrap";
 import { FacetFilterModel, FacetModel } from "../../../models/facets";
 import { SearchResponseModel } from "../../../models/dataset";
 import { useNavigate } from "react-router-dom";
-import { handleFilterAndSearch } from "../../../utils/utils";
 import { querySearchService } from "../../../api/browse";
+import { handleFilterAndSearch } from "../shared";
 
 interface DatasetHeaderProps {
   dsCount: number;
